@@ -37,53 +37,54 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "17.0.2",
-		"@angular/common": "17.0.2",
-		"@angular/compiler": "17.0.2",
-		"@angular/core": "17.0.2",
-		"@angular/forms": "17.0.2",
-		"@angular/platform-browser": "17.0.2",
-		"@angular/platform-browser-dynamic": "17.0.2",
-		"@angular/router": "17.0.2",
-		"@nx/angular": "17.1.2",
+		"@angular/animations": "18.0.5",
+		"@angular/common": "18.0.5",
+		"@angular/compiler": "18.0.5",
+		"@angular/core": "18.0.5",
+		"@angular/forms": "18.0.5",
+		"@angular/platform-browser": "18.0.5",
+		"@angular/platform-browser-dynamic": "18.0.5",
+		"@angular/router": "18.0.5",
+		"@ngneat/lib": "^3.0.1",
+		"@nx/angular": "19.3.2",
+		"angular-svg-icon": "^16.0.0",
 		"rxjs": "~7.8.0",
 		"tslib": "^2.3.0",
-		"zone.js": "0.14.2",
-		"@ngneat/lib": "^3.0.1",
-		"angular-svg-icon": "^16.0.0"
+		"zone.js": "0.14.7"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "17.0.0",
-		"@angular-devkit/core": "17.0.0",
-		"@angular-devkit/schematics": "17.0.0",
-		"@angular-eslint/eslint-plugin": "17.0.1",
-		"@angular-eslint/eslint-plugin-template": "17.0.1",
-		"@angular-eslint/template-parser": "17.0.1",
-		"@angular/cli": "~17.0.0",
-		"@angular/compiler-cli": "17.0.2",
-		"@angular/language-service": "17.0.2",
-		"@nx/cypress": "17.1.2",
-		"@nx/eslint-plugin": "17.1.2",
-		"@nx/jest": "17.1.2",
-		"@nx/js": "17.1.2",
-		"@nx/web": "17.1.2",
-		"@nx/workspace": "17.1.2",
-		"@schematics/angular": "17.0.0",
+		"@angular-devkit/build-angular": "18.0.6",
+		"@angular-devkit/core": "18.0.6",
+		"@angular-devkit/schematics": "18.0.6",
+		"@angular-eslint/eslint-plugin": "17.3.0",
+		"@angular-eslint/eslint-plugin-template": "17.3.0",
+		"@angular-eslint/template-parser": "17.3.0",
+		"@angular/cli": "~18.0.0",
+		"@angular/compiler-cli": "18.0.5",
+		"@angular/language-service": "18.0.5",
+		"@nx/cypress": "19.3.2",
+		"@nx/eslint": "19.3.2",
+		"@nx/eslint-plugin": "19.3.2",
+		"@nx/jest": "19.3.2",
+		"@nx/js": "19.3.2",
+		"@nx/web": "19.3.2",
+		"@nx/workspace": "19.3.2",
+		"@schematics/angular": "18.0.6",
 		"@types/jest": "^29.4.0",
-		"@types/node": "16.11.7",
+		"@types/node": "18.16.9",
 		"@typescript-eslint/eslint-plugin": "6.11.0",
 		"@typescript-eslint/parser": "6.11.0",
 		"autoprefixer": "^10.4.0",
-		"cypress": "^13.0.0",
+		"cypress": "^13.6.6",
 		"eslint": "8.46.0",
 		"eslint-config-prettier": "9.0.0",
-		"eslint-plugin-cypress": "2.15.1",
+		"eslint-plugin-cypress": "2.15.2",
+		"git-cz": "^3.2.1",
 		"jest": "^29.4.1",
 		"jest-environment-jsdom": "^29.4.1",
-		"jest-preset-angular": "13.1.4",
-		"ng-packagr": "17.0.0",
-		"nx": "17.1.2",
-		"git-cz": "^3.2.1",
+		"jest-preset-angular": "14.1.1",
+		"ng-packagr": "18.0.0",
+		"nx": "19.3.2",
 		"postcss": "^8.3.9",
 		"postcss-import": "14.1.0",
 		"postcss-preset-env": "7.5.0",
@@ -92,7 +93,6 @@
 		"standard-version": "^9.3.2",
 		"ts-jest": "29.1.0",
 		"ts-node": "10.9.1",
-		"typescript": "5.2.2",
-		"@nx/eslint": "17.1.2"
+		"typescript": "5.4.5"
 	}
 }
